Add ignore XFrame extensions for capture mode
-Extension loading does NOT work on firefox as some error is preventing
geckodriver from opening the profile created by Selenium in a temporary
directory. Upgrading to Selenium 4 does not work for the current
configuration as the ServiceLoader cannot load the FirefoxServiceBuilder
with most of the common maven packing methods to create an executable
jar file resulting in a instant crash when using firefox.